% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_apt.R
\name{term_apt}
\alias{term_apt}
\title{Apt}
\usage{
term_apt(packages, ppa = NULL)
}
\arguments{
\item{packages}{Character vector.
Packages to install.}

\item{ppa}{Character vector.
Personal Package Archives (PPA) to add.}
}
\description{
Installs packages on Debian based Linux distributions.
}
\author{
Ivan Jacob Agaloos Pesigan
}
