% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_bash.R
\name{term_bash}
\alias{term_bash}
\title{Setup Bash dotfiles}
\usage{
term_bash(
  dir = Sys.getenv("HOME"),
  overwrite = FALSE,
  vars = NULL,
  conda = FALSE,
  conda_path = NULL
)
}
\arguments{
\item{dir}{Character string.
\verb{.bash*} directory.
Defaults to user's home directory.}

\item{overwrite}{Logical.
Overwrite existing \code{bash} dot files in \code{dir}.}

\item{vars}{Named character vector.
Variables to export.
(e.g., \code{vars = c(GITHUB_PAT = "token_here", TRAVIS_TOKEN = "token_here")}.}

\item{conda}{Logical.
Add Miniconda \code{PATH} to \code{.bashrc}.}

\item{conda_path}{Character string.
Miniconda \code{PATH}.
If unspecified,
defaults to
\code{{HOME}/.local/miniconda3/bin}.}
}
\description{
Creates the following files:
\code{.bashrc},
\code{.bash_aliases},
\code{.bash_profile},
\code{.bash_logout}
in \code{dir}.
}
\examples{
\dontrun{
term_bash(
  dir = getwd(),
  overwrite = TRUE,
  vars = c(
    GITHUB_PAT = "123456",
    TRAVIS_TOKEN = "123456"
  )
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
