% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_brew.R
\name{term_brew}
\alias{term_brew}
\title{Homebrew}
\usage{
term_brew(packages, cask = FALSE, tap = NULL)
}
\arguments{
\item{packages}{Character vector.
Packages to install.}

\item{cask}{Logical.
If \code{TRUE}, uses \verb{brew cask install ...}.
If \code{FALSE}, uses \verb{brew install ...}.}

\item{tap}{Character vector.
Add third-party repositories.}
}
\description{
Install \code{Homebrew} and install packages.
}
