% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_choco.R
\name{term_choco}
\alias{term_choco}
\title{Chocolatey}
\usage{
term_choco(packages)
}
\arguments{
\item{packages}{Character vector.
Packages to install.}
}
\description{
Installs packages on Windows using Chocolatey.
}
\author{
Ivan Jacob Agaloos Pesigan
}
