% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_conda.R
\name{term_conda}
\alias{term_conda}
\title{Setup Conda for Bash}
\usage{
term_conda(
  dir = Sys.getenv("HOME"),
  overwrite = FALSE,
  conda_path = NULL,
  auto_activate_base = FALSE
)
}
\arguments{
\item{dir}{Character string.
\code{.bashrc} directory.
Defaults to user's home directory.}

\item{overwrite}{Logical.
Overwrite existing \code{.bashrc} and \code{.condarc} files in \code{dir}.
If \code{TRUE} and if file/s exist/s,
the settings are appended to the existing file.
If \code{FALSE} and if file/s exist/s,
none of the settings are written on disk.}

\item{conda_path}{Character string.
Miniconda \code{PATH}.
If unspecified,
defaults to \code{{HOME}/.local/miniconda3}.}

\item{auto_activate_base}{Logical.
Auto activate base.}
}
\description{
Appends \code{Conda} settings to
\code{.bashrc} and \code{.condarc}.
}
\examples{
\dontrun{
term_conda(
  dir = getwd(),
  overwrite = TRUE,
  conda_path = file.path(
    Sys.getenv("HOME"),
    ".local",
    "miniconda3"
  ),
  auto_activate_base = TRUE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
