% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_git_config.R
\name{term_git_config}
\alias{term_git_config}
\title{Git config}
\usage{
term_git_config(
  name = "User Name",
  email = "email@email.com",
  editor = "vim",
  ignore = c("*~", ".*.swp", ".DS_Store", ".~lock.*"),
  msg = paste0("Subject line (try to keep under 50 characters)", "\\n\\n",
    "Multi-line description of commit,", "\\n", "feel free to be detailed.", "\\n\\n",
    "[Ticket: X]"),
  global = TRUE
)
}
\arguments{
\item{name}{Character string.
\code{Git} \code{user.name}.}

\item{email}{Character string.
\code{Git} \code{user.email}.}

\item{editor}{Character string.
\code{Git} \code{core.editor}.}

\item{ignore}{Character vector.
Patterns of file names and paths to ignore.}

\item{msg}{Character string.
\code{Git} \code{commit.template}.}

\item{global}{Logical.
If \code{TRUE},
uses the \code{--global} option.
If \code{FALSE},
uses the \code{--local} option.}
}
\description{
Sets up \code{Git} configurations
including
user name,
email,
editor,
files to ignore,
and
commit message.
}
\examples{
\dontrun{
term_git_config(
  name = "User Name",
  email = "email@email.com",
  global = TRUE
)
}
}
