% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_install.R
\name{term_install}
\alias{term_install}
\title{Install Terminal Tools}
\usage{
term_install(packages, ...)
}
\arguments{
\item{packages}{Character vector.
Packages to install.}

\item{...}{Arguments to pass to \code{\link{term_brew}}.}
}
\description{
Installs terminal tools.
Uses package managers to install packages
(\code{pacman} for Arch Linux based Linux Distributions,
\code{apt} for Debian based Linux Distributions,
\code{hombrew} for Mac OSX and
\code{chocolatey} for Windows.)
}
\details{
NOTE: \code{chocolatey} for Windows is not currently implemented.
}
\examples{
\dontrun{
term_install()
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
