% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_pacman.R
\name{term_pacman}
\alias{term_pacman}
\title{Pacman}
\usage{
term_pacman(packages)
}
\arguments{
\item{packages}{Character vector.
Packages to install.}
}
\description{
Installs packages on Arch Linux based Linux distributions.
}
\author{
Ivan Jacob Agaloos Pesigan
}
