% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_reset_user_lib.R
\name{term_reset_user_lib}
\alias{term_reset_user_lib}
\title{Reset R User Library.}
\usage{
term_reset_user_lib()
}
\description{
Resets the R user library by doing the following:
\itemize{
\item removes contents of R user library (\code{{HOME}/R/{PLATFORM}/{R.VERSION}}) and
\item executes \code{term_user_lib} with arguments \code{libpath = NULL},
\code{dir = Sys.getenv("HOME")}, and \code{overwrite = TRUE}.
\strong{WARNING:}
This function
\strong{permanently deletes files in \code{{HOME}/R/{PLATFORM}/{R.VERSION}}}.
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
