% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_vim.R
\name{term_vim}
\alias{term_vim}
\title{Create \code{.vimrc} File and Install Plugins}
\usage{
term_vim(dir = Sys.getenv("HOME"), overwrite = FALSE, plugins = FALSE)
}
\arguments{
\item{dir}{Character string.
\code{.vimrc} directory}

\item{overwrite}{Logical.
Overwrite existing \code{.vimrc} file in \code{dir}.}

\item{plugins}{Logical.
Install \code{vim} plugins.}
}
\description{
Creates a \code{.vimrc} file in \code{dir}
and installs plugins.
}
\examples{
\dontrun{
term_vim(
  dir = getwd(),
  overwrite = TRUE,
  plugins = TRUE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
