% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_reset_user_lib.R
\name{term_reset_user_lib}
\alias{term_reset_user_lib}
\title{Reset R User Library.}
\usage{
term_reset_user_lib(libpath = NULL, dir = Sys.getenv("HOME"), overwrite = TRUE)
}
\arguments{
\item{libpath}{Character string.
User library path.
If unspecified, defaults to
\code{{HOME}/R/{PLATFORM}/{R.VERSION}}
(\verb{$\{HOME\}/R/\%p/\%v}).}

\item{dir}{Character string.
\code{.Renviron} directory.
Defaults to user's home directory.}

\item{overwrite}{Logical.
If \code{.Renviron} exists in \code{dir},
the variable \code{R_LIBS_USER}
is overwritten
with  \code{R_LIBS_USER="{libpath}"}.
This argument is set to \code{TRUE}
to ensure that the specified \code{libpath}
is reflected in the \code{.Renviron} in \code{dir}.}
}
\description{
Resets the R user library by doing the following:
\itemize{
\item removes contents of R user library (\code{libpath}) and
\item executes \code{term_user_lib}
\strong{WARNING:}
This function
\strong{permanently deletes files in \code{libpath}}.
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
