% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_xterm.R
\name{term_xterm}
\alias{term_xterm}
\title{Create Xterm .Xresources}
\usage{
term_xterm(dir = Sys.getenv("HOME"), overwrite = FALSE, dark = TRUE)
}
\arguments{
\item{dir}{Character string.
\code{.Xresources} directory.
Defaults to user's home directory.}

\item{overwrite}{Logical.
Overwrite existing \code{.Xresources} dot files in the home directory.}

\item{dark}{Logical.
Dark color scheme.}
}
\description{
Creates \code{Xterm} \code{.Xresources} in the home directory.
}
\examples{
\dontrun{
term_xterm(
  overwrite = TRUE,
  dark = TRUE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
